Library

Library

Public API

Generating Lifting Surfaces

Manipulating Lifting Surfaces

Reference Parameters and Frames

Freestream Parameters

Solving for Circulation

Near Field Forces and Moments

Far Field Drag

Body and Stability Derivatives

Visualization

Private API

Geometry

Vortex Lattice Panels

Freestream

Circulation

Near-Field

Far-Field

Index

Public API

Generating Lifting Surfaces

VortexLattice.AbstractSpacing — Type

AbstractSpacing

Spacing discretization scheme supertype

VortexLattice.Uniform — Type

Uniform()

Uniform discretization scheme

VortexLattice.Sine — Type

Sine()

Sine-spaced discretization scheme.

VortexLattice.Cosine — Type

Cosine()

Cosine-spaced discretization scheme

VortexLattice.grid_to_horseshoe_vortices — Method

grid_to_horseshoe_vortices(xyz; mirror=false, fcore=(c, Δs)->max(c/4, Δs))

Construct a set of horseshoe vortex panels given a potentially curved lifting surface defined by a grid with dimensions (3, i, j) where i corresponds to the chordwise direction (ordered from leading edge to trailing edge) and j corresponds to the spanwise direction (ordered from left to right). The bound vortex will be placed at the 1/4 chord and the control point will be placed at the 3/4 chord of each panel.

In order to be consistent with vortex lattice theory using horseshoe vortices, the y and z-coordinates of each resulting chordwise strip of panels are set to the same y and z-coordinates as found at the leading edge of the chordwise strip. The normal vectors used to satisfy the no-flow boundary conditions, however, are set using the original geometry.

The optional argument mirror may be used to mirror the geometry across the y-axis.

The argument fcore defines a function for setting the finite core size based on the chord length in the x-direction and/or the panel width in the y/z directions. The default function is the same as that used by AVL.

VortexLattice.grid_to_vortex_rings — Method

grid_to_vortex_rings(xyz; mirror=false, fcore=(c, Δs)->max(c/4, Δs/2))

Construct a set of vortex ring panels given a potentially curved lifting surface defined by a grid with dimensions (3, i, j) where i corresponds to the chordwise direction (ordered from leading edge to trailing edge) and j corresponds to the spanwise direction (ordered from left to right). The leading edge of each ring vortex will be placed at the 1/4 chord and the control point will be placed at the 3/4 chord of each panel.

The optional argument mirror may be used to mirror the geometry across the y-axis.

The argument fcore defines a function for setting the finite core size based on the chord length in the x-direction and/or the panel width in the y/z directions. The default function is the same as that used by AVL.

VortexLattice.grid_to_horseshoe_vortices — Method

grid_to_horseshoe_vortices(xyz, ns, nc; mirror=false,
    fcore=(c, Δs)->max(c/4, Δs/2), spacing_s=Cosine(), spacing_c=Uniform(),
    interp_s=(x)->(x,y,xpt)->LinearInterpolation(x, y)(xpt),
    interp_c=(x,y,xpt)->LinearInterpolation(x, y)(xpt))

Discretize a potentially curved lifting surface defined by a grid with dimensions (3, i, j) where i corresponds to the chordwise direction (ordered from leading edge to trailing edge) and j corresponds to the spanwise direction (ordered from left to right) into ns spanwise and nc chordwise horseshoe vortex panels according to the spanwise discretization scheme spacing_s and chordwise discretization scheme spacing_c. The bound vortex will be placed at the 1/4 chord and the control point will be placed at the 3/4 chord of each panel.

In order to be consistent with vortex lattice theory using horseshoe vortices, the y and z-coordinates of each resulting chordwise strip of panels are set to the same y and z-coordinates as found at the leading edge of the chordwise strip. The normal vectors used to satisfy the no-flow boundary conditions, however, are set using the original geometry.

Arguments

xyz: grid of dimensions (3, i, j) where where i corresponds to the chordwise direction and j corresponds to the spanwise direction.

ns: number of spanwise panels

nc: number of chordwise panels

mirror: mirror the geometry across the y-axis? defaults to false.

fcore: function for setting the finite core size based on the chord length (in the x-direction) and/or the panel width (in the y/z directions). Defaults to (c, Δs)->max(c/4, Δs/2) (which corresponds to AVL's implementation)

spacing_s: spanwise discretization scheme, defaults to Cosine()

spacing_c: chordwise discretization scheme, defaults to Uniform()

interp_s: spanwise interpolation function, defaults to linear interpolation

interp_c: chordwise interpolation function, defaults to linear interpolation

VortexLattice.grid_to_vortex_rings — Method

grid_to_vortex_rings(xyz, ns, nc; mirror=false, fcore=(c, Δs)->max(c/4, Δs/2),
    spacing_s=Cosine(), spacing_c=Uniform(),
    interp_s=(x,y,xpt)->LinearInterpolation(x, y)(xpt),
    interp_c=(x,y,xpt)->LinearInterpolation(x, y)(xpt))

Discretize a potentially curved lifting surface defined by a grid with dimensions (3, i, j) where i corresponds to the chordwise direction (ordered from leading edge to trailing edge) and j corresponds to the spanwise direction (ordered from left to right) into ns spanwise and nc chordwise vortex ring panels according to the spanwise discretization scheme spacing_s and chordwise discretization scheme spacing_c. The bound vortex will be placed at the 1/4 chord and the control point will be placed at the 3/4 chord of each panel

Arguments

xyz: grid of dimensions (3, i, j) where where i corresponds to the chordwise direction and j corresponds to the spanwise direction.

ns: number of spanwise panels

nc: number of chordwise panels

mirror: mirror the geometry across the y-axis? defaults to false.

fcore: function for setting the finite core size based on the chord length (in the x-direction) and/or the panel width (in the y/z directions). Defaults to (c, Δs)->max(c/4, Δs/2) (which corresponds to AVL's implementation)

spacing_s: spanwise discretization scheme, defaults to Cosine()

spacing_c: chordwise discretization scheme, defaults to Uniform()

interp_s: spanwise interpolation function, defaults to linear interpolation

interp_c: chordwise interpolation function, defaults to linear interpolation

VortexLattice.wing_to_horseshoe_vortices — Function

wing_to_horseshoe_vortices(xle, yle, zle, chord, theta, phi, ns, nc;
    fc=fill(x->0, length(xle)), mirror=false, fcore=(c, Δs)->max(c/4, Δs/2),
    spacing_s=Cosine(), spacing_c=Uniform(), interp_s=(x,y,xpt)->LinearInterpolation(x, y)(xpt))

Discretize a wing into ns spanwise and nc chordwise horseshoe vortex panels according to the spanwise discretization scheme spacing_s and chordwise discretization scheme spacing_c.

Arguments

xle: leading edge x-coordinate of each airfoil section

yle: leading edge y-coordinate of each airfoil section

zle: leading edge z-coordinate of each airfoil section

chord: chord length of each airfoil section

theta: twist of each airfoil section

phi: dihedral angle of each airfoil section, defined by a right hand rotation about the x-axis

ns: number of spanwise panels

nc: number of chordwise panels

fc: (optional) camber line function y=f(x) of each airfoil section

mirror: mirror the geometry across the y-axis?, defaults to false

fcore: function for setting the finite core size based on the chord length (in the x-direction) and/or the panel width (in the y/z directions). Defaults to (c, Δs)->max(c/4, Δs/2) (which corresponds to AVL's implementation)

spacing_s: spanwise discretization scheme, defaults to Cosine()

spacing_c: chordwise discretization scheme, defaults to Uniform()

interp_s: interpolation function between spanwise stations, defaults to linear interpolation

VortexLattice.wing_to_vortex_rings — Function

wing_to_vortex_rings(xle, yle, zle, chord, theta, phi, ns, nc;
    fc=fill(x->0, length(xle)), mirror=false,
    fcore=(c, Δs)->max(c/4, Δs/2), spacing_s=Cosine(), spacing_c=Uniform(),
    interp_s=(x,y,xpt)->LinearInterpolation(x, y)(xpt))

Discretize a wing into ns spanwise and nc chordwise vortex ring panels according to the spanwise discretization scheme spacing_s and chordwise discretization scheme spacing_c.

Arguments

xle: leading edge x-coordinate of each airfoil section

yle: leading edge y-coordinate of each airfoil section

zle: leading edge z-coordinate of each airfoil section

chord: chord length of each airfoil section

theta: twist of each airfoil section

phi: dihedral angle of each airfoil section, defined by a right hand rotation about the x-axis

ns: number of spanwise panels

nc: number of chordwise panels

fc: (optional) camber line function y=f(x) of each airfoil section

mirror: mirror the geometry across the y-axis?, defaults to false

fcore: function for setting the finite core size based on the chord length (in the x-direction) and/or the panel width (in the y/z directions). Defaults to (c, Δs)->max(c/4, Δs/2) (which corresponds to AVL's implementation)

spacing_s: spanwise discretization scheme, defaults to Cosine()

spacing_c: chordwise discretization scheme, defaults to Uniform()

interp_s: interpolation function between spanwise stations, defaults to linear interpolation

Manipulating Lifting Surfaces

VortexLattice.AbstractPanel — Type

AbstractPanel{TF}

Supertype of vortex lattice method panel types.

VortexLattice.Horseshoe — Type

Horseshoe{TF}

Horseshoe shaped panel element with trailing vortices that extend in the +x direction to the trailing edge, and then into the farfield.

Fields

rl: position of the left side of the bound vortex

rc: position of the center of the bound vortex

rr: position of the right side of the bound vortex

rcp: position of the panel control point

ncp: normal vector at the panel control point

xl_te: x-distance from the left side of the bound vortex to the trailing edge

xc_te: x-distance from the center of the bound vortex to the trailing edge

xr_te: x-distance from the right side of the bound vortex to the trailing edge

core_size: finite core size

VortexLattice.Ring — Type

Ring{TF}

Vortex ring shaped panel element.

Fields

rtl: position of the left side of the top bound vortex

rtc: position of the center of the top bound vortex

rtr: position of the right side of the top bound vortex

rbl: position of the left side of the bottom bound vortex

rbc: position of the center of the bottom bound vortex

rbr: position of the right side of the bottom bound vortex

rcp: position of the panel control point

ncp: normal vector at the panel control point

core_size: finite core size

VortexLattice.translate — Method

translate(panels::AbstractVector{<:AbstractPanel}, r)

Return a copy of panels translated the distance specified by vector r

VortexLattice.translate! — Method

translate!(panels, r)

Translate the panels contained in panels the distance specified by vector r

VortexLattice.reflect — Method

reflect(panels::AbstractMatrix{<:AbstractPanel})

Reflects panels about the y-axis, preserving panel grid-based ordering

Reference Parameters and Frames

VortexLattice.Reference — Type

Reference(S, c, b, r)

Reference quantities.

Arguments

S: reference area

c: reference chord

b: reference span

r: reference location for all rotations/moments

Freestream Parameters

VortexLattice.Freestream — Type

Freestream(alpha, beta, Omega, additional_velocity=nothing)

Define the freestream properties.

Arguments

alpha: angle of attack (rad)

beta: sideslip angle (rad)

Omega: rotation vector (p, q, r) of the body frame about the center of gravity, normalized by Vinf

additional_velocity: a function of the form: V = additional_velocity(r) which returns the additional velocity V (normalized by the freestream velocity) at position r. Defaults to nothing.

VortexLattice.body_to_stability — Function

body_to_stability(fs::Freestream)

Construct a rotation matrix from the body axis to the stability axis.

VortexLattice.body_to_wind — Function

body_to_wind(fs::Freestream)

Construct a rotation matrix from the body axis to the wind axis

VortexLattice.stability_to_body — Function

stability_to_body(fs::Freestream)

Construct a rotation matrix from the stability axis to the body axis

VortexLattice.stability_to_wind — Function

stability_to_wind(fs::Freestream)

Construct a rotation matrix from the stability axis to the wind axis

VortexLattice.wind_to_body — Function

wind_to_body(fs::Freestream)

Construct a rotation matrix from the wind axis to the body axis

VortexLattice.wind_to_stability — Function

wind_to_stability(fs::Freestream)

Construct a rotation matrix from the wind axis to the stability axis

Solving for Circulation

VortexLattice.influence_coefficients — Method

influence_coefficients(surface, symmetric; kwargs...)

Construct the aerodynamic influence coefficient matrix for a single surface.

Arguments:

surface: Matrix of panels of shape (nc, ns) where nc is the number of chordwise panels and ns is the number of spanwise panels

symmetric: Flag indicating whether a mirror image of the panels in surface should be used when calculating induced velocities.

xhat: (optional) direction in which trailing vortices are shed, defaults to [1, 0, 0]

VortexLattice.influence_coefficients — Method

influence_coefficients(surfaces, surface_id, symmetric; kwargs...)

Construct the aerodynamic influence coefficient matrix for multiple surfaces.

Arguments:

surface: Vector of surfaces, represented by matrices of panels of shape (nc, ns) where nc is the number of chordwise panels and ns is the number of spanwise panels

surface_id: ID for each surface. May be used to deactivate the finite core model by setting all surface ID's to the same value.

symmetric: Flag indicating whether a mirror image of the panels in surface should be used when calculating induced velocities.

xhat: (optional) direction in which trailing vortices are shed, defaults to [1, 0, 0]

VortexLattice.influence_coefficients! — Method

influence_coefficients!(AIC, surface, symmetric; kwargs...)

Pre-allocated version of influence_coefficients

VortexLattice.influence_coefficients! — Method

influence_coefficients!(AIC, surfaces, surface_id, symmetric; kwargs...)

Pre-allocated version of influence_coefficients

VortexLattice.normal_velocity — Function

normal_velocity(surface[s], ref, fs)

Compute the normal component of the external velocity for a single surface or for a vector of surfaces.

This forms the right hand side of the circulation linear system solve.

VortexLattice.normal_velocity! — Function

normal_velocity!(b, surface[s], ref, fs)

Non-allocating version of normal_velocity

VortexLattice.circulation — Function

circulation(AIC, b)

Solve for the circulation distribution.

Near Field Forces and Moments

VortexLattice.AbstractFrame — Type

AbstractFrame

Supertype for the different possible reference frames used by this package.

VortexLattice.Body — Type

Body <: AbstractFrame

Reference frame aligned with the global X-Y-Z axes

VortexLattice.Stability — Type

Stability <: AbstractFrame

Reference frame rotated from the body frame about the y-axis to be aligned with the freestream alpha.

VortexLattice.Wind — Type

Wind <: AbstractFrame

Reference frame rotated to be aligned with the freestream alpha and beta

VortexLattice.PanelProperties — Type

PanelProperties

Panel specific properties calculated during the vortex lattice method analysis.

Fields

gamma: Panel circulation strength (normalized by the freestream velocity)

v: Local velocity at the panel's center (typically the quarter-chord), normalized

cf: Bound vortex force per unit length, normalized by QINF*S where QINF is the dynamic pressure and S is the user-provided reference area.

cfl: Left vortex force per unit length, normalized by QINF*S

cfr: Right vortex force per unit length, normalized by QINF*S

VortexLattice.near_field_forces — Method

near_field_forces(surface, reference, freestream, symmetric, Γ; kwargs...)

Compute the forces and moments acting on the aircraft given the circulation distribution Γ.

Return CF, CM, and a vector of panel properties of type PanelProperties.

Arguments:

surface: Matrix of panels of shape (nc, ns) where nc is the number of chordwise panels and ns is the number of spanwise panels

reference: reference parameters (see Reference)

freestream: freestream parameters (see Freestream)

symmetric: Flag indicating whether a mirror image of the panels in surface should be used when calculating induced velocities.

Γ: Vector of circulation strengths for each panel, calculated by circulation

xhat: (optional) direction in which trailing vortices are shed, defaults to [1, 0, 0]

frame: frame in which to return CF and CM, options are Body() (default), Stability(), and Wind()

VortexLattice.near_field_forces — Method

near_field_forces(surfaces, surface_id, reference, freestream, symmetric, Γ; kwargs...)

Compute the forces and moments acting on the aircraft given the circulation distribution Γ.

Return CF, CM, and a vector of panel properties of type PanelProperties.

Arguments:

surface: Vector of surfaces, represented by matrices of panels of shape (nc, ns) where nc is the number of chordwise panels and ns is the number of spanwise panels

surface_id: ID for each surface. May be used to deactivate the finite core model by setting all surface ID's to the same value.

reference: reference parameters (see Reference)

freestream: freestream parameters (see Freestream)

symmetric: Flag indicating whether a mirror image of the panels in surface should be used when calculating induced velocities.

Γ: Vector of circulation strengths for each panel, calculated by circulation

xhat: (optional) direction in which trailing vortices are shed, defaults to [1, 0, 0]

frame: frame in which to return CF and CM, options are Body() (default), Stability(), and Wind()

Far Field Drag

VortexLattice.far_field_drag — Function

far_field_drag(surface[s], ref, fs, symmetric, Γ)

Computes induced drag using the Trefftz plane (far field method).

Body and Stability Derivatives

VortexLattice.body_derivatives — Function

body_derivatives(panels, ref, fs, symmetric, AIC; xhat=[1,0,0])

Returns the derivatives of the body forces and moments with respect to the freestream velocity components (u, v, w) and the angular velocity components (p, q, r) in the body frame.

The derivatives are returned as two named tuples: dCF, dCM

VortexLattice.stability_derivatives — Function

stability_derivatives(panels, ref, fs, symmetric, AIC; xhat=[1,0,0])

Returns the derivatives of the body forces and moments in the stability frame with respect to the freestream velocity components (alpha, beta) and the angular velocity components (p, q, r) in the stability frame.

The derivatives are returned as two named tuples: dCF, dCM

Visualization

VortexLattice.write_vtk — Function

write_vtk(name, panels::AbstractVector{<:Horseshoe}; kwargs...)

Writes the geometry to Paraview files for visualization.

Keyword Arguments:

xhat=[1, 0, 0]: direction in which trailing vortices are shed

mirror=false: creates a mirror image of the geometry across the X-Z axis

wake_length=10: distance to extend the trailing vortices

metadata=Dict(): dictionary of metadata to include in generated files

Private API

Geometry

VortexLattice.linearinterp — Function

linearinterp(eta, rstart, rend)

Linearly interpolate between rstart and rend where eta is the fraction between 0 (rstart) and 1 (rend)

VortexLattice.spanwise_spacing — Function

spanwise_spacing(n, spacing::AbstractSpacing)

Distribute n panel endpoints and n-1 panel midpoints on the interval between 0 and 1 according to the discretization strategy in spacing.

VortexLattice.chordwise_spacing — Function

spanwise_spacing(n, spacing::AbstractSpacing)

Distribute n-1 vortex and n-1 control point chordwise locations on the interval between 0 and 1 according to the discretization strategy in spacing.

VortexLattice.interpolate_grid — Function

interpolate_grid(xyz, eta, interp; xdir=0, ydir=1)

Interpolates the grid xyz along direction dir

Arguments

xyz: Grid of size (3, ni, nj)

eta: New (normalized) coordinates in direction dir (0 <= eta <= 1)

interp: Interpolation method of form ypt = f(x,y,xpt)

xdir: Independent variable direction, defaults to arc length

ydir: Dependent variable direction xyz (i=1, j=2)

Vortex Lattice Panels

VortexLattice.top_left — Function

top_left(panel::AbstractPanel)

Return the top left vertex of panel

VortexLattice.top_center — Function

top_center(panel::AbstractPanel)

Return the top center vertex of panel

VortexLattice.top_right — Function

top_right(panel::AbstractPanel)

Return the top right vertex of panel

VortexLattice.bottom_left — Function

bottom_left(panel::AbstractPanel)

Return the bottom left vertex of panel

VortexLattice.bottom_center — Function

bottom_center(panel::AbstractPanel)

Return the bottom center vertex of panel

VortexLattice.bottom_right — Function

bottom_right(panel::AbstractPanel)

Return the bottom right vertex of panel

VortexLattice.controlpoint — Function

controlpoint(panel::AbstractPanel)

Return the control point of panel (typically located at the 3/4 chord)

VortexLattice.normal — Function

normal(panel::AbstractPanel)

Return the normal vector of panel

VortexLattice.get_core_size — Function

get_core_size(panel::AbstractPanel)

Return the panel core size

VortexLattice.translate — Method

translate(panel::AbstractPanel, r)

Return a copy of panel translated the distance specified by vector r

VortexLattice.reflect — Method

reflect(panel::AbstractPanel)

Reflects a panel across the y-axis.

VortexLattice.induced_velocity — Function

induced_velocity(rcp, panel, same_id, symmetric, xhat, args...)

Computes the normalized induced velocity at control point rcp from panel.

Arguments

rcp: Control point for receiving panel

panel: Sending panel

same_id: Flag indicating whether the sending and receiving surfaces have the same surface ID

symmetric: Flag indicating whether sending panel should be mirrored across the y-axis

xhat: Direction in which to shed trailing vortices

Optional Arguments for Horseshoe Vortices

include_bound: Indicates whether to include the bound vortex in induced velocity calculations, defaults to true

Optional Arguments for Vortex Rings

trailing: Flag indicating whether sending panel sheds trailing vortices

include_top: Indicates whether to include the top bound vortex in induced velocity calculations, defaults to true

include_bottom: Indicates whether to include the bottom bound vortex in induced velocity calculations, defaults to true

VortexLattice.panel_induced_velocity — Function

panel_induced_velocity(receiving, Ir, sending, Is, same_surface, same_id, symmetric, xhat, trailing)

Computes the normalized induced velocity on panel receiving from panel sending.

Arguments

receiving: Receiving panel

Ir: Indices of receiving panel

sending: Sending panel

Is: Indices of sending panel

same_surface: Flag indicating whether the sending and receiving surfaces are the same

same_id: Flag indicating whether the sending and receiving surfaces have the same surface ID

symmetric: Flag indicating whether sending panel should be mirrored across the y-axis

xhat: Direction in which to shed trailing vortices

trailing: Indicates whether the panel is on the trailing edge

VortexLattice.panel_circulation — Function

panel_circulation(panel, Γ1, Γ2)

Return the circulation on panel given the circulation strength of the previous bound vortex Γ1 and the current bound vortex Γ2

VortexLattice.influence_coefficients! — Method

influence_coefficients!(AIC, receiving::AbstractMatrix{<:AbstractPanel},
    sending::AbstractMatrix{<:AbstractPanel}, same_id, symmetric, xhat)

Compute the AIC coefficients corresponding to the influence of the panels in sending on the panels in receiving.

VortexLattice.left_center — Function

left_center(panel::AbstractPanel)

Return the center of the left bound vortex on panel

VortexLattice.right_center — Function

right_center(panel::AbstractPanel)

Return the center of the right bound vortex on panel

VortexLattice.top_vector — Function

top_vector(panel::AbstractPanel)

Returns the path of the top bound vortex for panel

VortexLattice.left_vector — Function

left_vector(panel)

Return the path of the left bound vortex for panel

VortexLattice.right_vector — Function

right_vector(panel)

Return the path of the right bound vortex for panel

VortexLattice.bottom_vector — Function

bottom_vector(panel)

Return the path of the bottom bound vortex for panel

VortexLattice.flipy — Function

flipy(r)

Flip sign of y-component of vector r (used for symmetry)

VortexLattice.not_on_symmetry_plane — Function

not_on_symmetry_plane(r1, r2, tol=eps())

Test whether points r1 and r2 are on the symmetry plane (y = 0)

VortexLattice.trailing_induced_velocity — Function

trailing_induced_velocity(r1, r2, xhat, same_id, core_size)

Compute the induced velocity (per unit circulation) for two vortices trailing in the xhat direction, at a control point located at r1 relative to the start of the left trailing vortex and r2 relative to the start of the right trailing vortex.

VortexLattice.bound_induced_velocity — Function

bound_induced_velocity(r1, r2, same_id, core_size)

Compute the induced velocity (per unit circulation) for a bound vortex, at a control point located at r1 relative to the start of the bound vortex and r2 relative to the end of the bound vortex

Freestream

VortexLattice.body_to_stability_alpha — Function

body_to_stability_alpha(fs::Freestream)

Construct a rotation matrix from the body axis to the stability axis and its derivative with respect to alpha

VortexLattice.body_to_wind_derivatives — Function

body_to_wind_derivatives(fs::Freestream)

Construct a rotation matrix from the body axis to the wind axis and its derivatives with respect to alpha and beta

VortexLattice.stability_to_body_alpha — Function

stability_to_body(fs::Freestream)

Construct a rotation matrix from the stability axis to the body axis and its derivative with respect to alpha

VortexLattice.stability_to_wind_beta — Function

stability_to_wind_beta(fs::Freestream)

Construct a rotation matrix from the stability axis to the wind axis and its derivative with respect to beta

VortexLattice.wind_to_body_derivatives — Function

wind_to_body_derivatives(fs::Freestream)

Construct a rotation matrix from the wind axis to the body axis and its derivatives with respect to alpha and beta

VortexLattice.wind_to_stability_beta — Function

wind_to_stability_beta(fs::Freestream)

Construct a rotation matrix from the wind axis to the stability axis and its derivative with respect to beta

VortexLattice.freestream_velocity — Function

freestream_velocity(freestream)

Computes the freestream velocity

VortexLattice.freestream_velocity_derivatives — Function

freestream_velocity_derivatives(freestream)

Computes the freestream velocity

VortexLattice.external_velocity — Function

external_velocity(freestream, r, rref)

Compute the external velocity at location r

VortexLattice.external_velocity_derivatives — Function

external_velocity_derivatives(freestream, r, rref)

Compute the external velocity at location r and its derivatives with respect to (alpha, beta, p, q, r)

Circulation

VortexLattice.normal_velocity_derivatives — Function

normal_velocity_derivatives(surface[s], ref, fs)

Compute the normal component of the external velocity for a single surface or for a vector of surfaces and its derivatives with respect to (alpha, beta, p, q, r).

This forms the right hand side of the circulation linear system solve (and its derivatives).

VortexLattice.normal_velocity_derivatives! — Function

normal_velocity_derivatives!(b, db, surface[s], ref, fs)

Non-allocating version of normal_velocity_derivatives

VortexLattice.circulation_derivatives — Function

circulation_derivatives(AIC, b, db)

Solve for the circulation distribution and its derivatives with respect to (alpha, beta, p, q, r) .

Near-Field

VortexLattice.body_to_frame — Function

body_to_frame(CF, CM, ref, fs, frame)

Transforms the coefficients CF and CM which are provided in the body frame to another frame

VortexLattice.near_field_forces_derivatives — Method

near_field_forces_derivatives(surface, reference, freestream, symmetric, Γ, dΓ; kwargs...)

Compute the forces and moments acting on the aircraft given the circulation distribution Γ and their derivatives with respect to the variables in freestream.

Return CF, CM, dCF, dCM, and a vector of panel properties of type PanelProperties. CF, CM, dCF, and dCM are returned in the body frame.

Arguments:

surface: Matrix of panels of shape (nc, ns) where nc is the number of chordwise panels and ns is the number of spanwise panels

reference: reference parameters (see Reference)

freestream: freestream parameters (see Freestream)

symmetric: Flag indicating whether a mirror image of the panels in surface should be used when calculating induced velocities.

Γ: Vector of circulation strengths for each panel

dΓ: Tuple containing the derivative of the circulation strength with respect to each of the freestream parameters

xhat: (optional) direction in which trailing vortices are shed, defaults to [1, 0, 0]

VortexLattice.near_field_forces_derivatives — Method

near_field_forces_derivatives(surfaces, surface_id, reference, freestream, symmetric, Γ; kwargs...)

Compute the forces and moments acting on the aircraft given the circulation distribution Γ and their derivatives with respect to the variables in freestream.

Return CF, CM, dCF, dCM, and a vector of panel properties of type PanelProperties. CF, CM, dCF, and dCM are returned in the body frame.

Arguments:

surface: Vector of surfaces, represented by matrices of panels of shape (nc, ns) where nc is the number of chordwise panels and ns is the number of spanwise panels

surface_id: ID for each surface. May be used to deactivate the finite core model by setting all surface ID's to the same value.

reference: reference parameters (see Reference)

freestream: freestream parameters (see Freestream)

symmetric: Flag indicating whether a mirror image of the panels in surface should be used when calculating induced velocities.

Γ: Vector of circulation strengths for each panel

dΓ: Tuple containing the derivative of the circulation strength with respect to each of the freestream parameters

xhat: (optional) direction in which trailing vortices are shed, defaults to [1, 0, 0]

Far-Field

VortexLattice.TrefftzPanel — Type

TrefftzPanel{TF}

Panel in the Trefftz plane.

VortexLattice.normal — Method

normal(panel::TrefftzPanel)

Return the normal vector of panel, including magnitude

VortexLattice.trefftz_panels — Function

trefftz_panels(surface[s], freestream, Γ)

Constructs a set of panels for Trefftz plane calculations

VortexLattice.panel_induced_drag — Function

panel_induced_drag(receiving::TrefftzPanel, sending::TrefftzPanel, same_id, symmetric)

Induced drag on receiving panel induced by sending panel.

VortexLattice.vortex_induced_drag — Function

vortex_induced_drag(rj, Γj, ri, Γi, nhati)

Return induced drag from vortex j induced on panel i

Index

VortexLattice.AbstractFrame

VortexLattice.AbstractPanel

VortexLattice.AbstractSpacing

VortexLattice.Body

VortexLattice.Cosine

VortexLattice.Freestream

VortexLattice.Horseshoe

VortexLattice.PanelProperties

VortexLattice.Reference

VortexLattice.Ring

VortexLattice.Sine

VortexLattice.Stability

VortexLattice.TrefftzPanel

VortexLattice.Uniform

VortexLattice.Wind

VortexLattice.body_derivatives

VortexLattice.body_to_frame

VortexLattice.body_to_stability

VortexLattice.body_to_stability_alpha

VortexLattice.body_to_wind

VortexLattice.body_to_wind_derivatives

VortexLattice.bottom_center

VortexLattice.bottom_left

VortexLattice.bottom_right

VortexLattice.bottom_vector

VortexLattice.bound_induced_velocity

VortexLattice.chordwise_spacing

VortexLattice.circulation

VortexLattice.circulation_derivatives

VortexLattice.controlpoint

VortexLattice.external_velocity

VortexLattice.external_velocity_derivatives

VortexLattice.far_field_drag

VortexLattice.flipy

VortexLattice.freestream_velocity

VortexLattice.freestream_velocity_derivatives

VortexLattice.get_core_size

VortexLattice.grid_to_horseshoe_vortices

VortexLattice.grid_to_horseshoe_vortices

VortexLattice.grid_to_vortex_rings

VortexLattice.grid_to_vortex_rings

VortexLattice.induced_velocity

VortexLattice.influence_coefficients

VortexLattice.influence_coefficients

VortexLattice.influence_coefficients!

VortexLattice.influence_coefficients!

VortexLattice.influence_coefficients!

VortexLattice.interpolate_grid

VortexLattice.left_center

VortexLattice.left_vector

VortexLattice.linearinterp

VortexLattice.near_field_forces

VortexLattice.near_field_forces

VortexLattice.near_field_forces_derivatives

VortexLattice.near_field_forces_derivatives

VortexLattice.normal

VortexLattice.normal

VortexLattice.normal_velocity

VortexLattice.normal_velocity!

VortexLattice.normal_velocity_derivatives

VortexLattice.normal_velocity_derivatives!

VortexLattice.not_on_symmetry_plane

VortexLattice.panel_circulation

VortexLattice.panel_induced_drag

VortexLattice.panel_induced_velocity

VortexLattice.reflect

VortexLattice.reflect

VortexLattice.right_center

VortexLattice.right_vector

VortexLattice.spanwise_spacing

VortexLattice.stability_derivatives

VortexLattice.stability_to_body

VortexLattice.stability_to_body_alpha

VortexLattice.stability_to_wind

VortexLattice.stability_to_wind_beta

VortexLattice.top_center

VortexLattice.top_left

VortexLattice.top_right

VortexLattice.top_vector

VortexLattice.trailing_induced_velocity

VortexLattice.translate

VortexLattice.translate

VortexLattice.translate!

VortexLattice.trefftz_panels

VortexLattice.vortex_induced_drag

VortexLattice.wind_to_body

VortexLattice.wind_to_body_derivatives

VortexLattice.wind_to_stability

VortexLattice.wind_to_stability_beta

VortexLattice.wing_to_horseshoe_vortices

VortexLattice.wing_to_vortex_rings

VortexLattice.write_vtk